# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ApplicationDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "active_analysis_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Analysis Id"
          },
          "active_job_snapshot_id": {
            "title": "Active Job Snapshot Id",
            "type": "string"
          },
          "active_operation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OperationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_selection_plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Selection Plan Id"
          },
          "active_working_draft_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Working Draft Id"
          },
          "application": {
            "$ref": "#/components/schemas/ApplicationResponse"
          },
          "available_actions": {
            "items": {
              "type": "string"
            },
            "title": "Available Actions",
            "type": "array"
          },
          "blocked_actions": {
            "items": {
              "$ref": "#/components/schemas/BlockedActionResponse"
            },
            "title": "Blocked Actions",
            "type": "array"
          },
          "latest_analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobAnalysisResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "latest_approved_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Approved Revision Id"
          },
          "latest_ready_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Ready Revision Id"
          },
          "latest_snapshot": {
            "$ref": "#/components/schemas/JobSnapshotResponse"
          },
          "newer_draft_in_progress": {
            "title": "Newer Draft In Progress",
            "type": "boolean"
          },
          "preparation_state": {
            "$ref": "#/components/schemas/PreparationState"
          },
          "primary_stale_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Stale Reason"
          },
          "recommended_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Action"
          },
          "recruitment_status": {
            "title": "Recruitment Status",
            "type": "string"
          },
          "review_reasons": {
            "items": {
              "$ref": "#/components/schemas/ReasonResponse"
            },
            "title": "Review Reasons",
            "type": "array"
          },
          "stale_reasons": {
            "items": {
              "$ref": "#/components/schemas/ReasonResponse"
            },
            "title": "Stale Reasons",
            "type": "array"
          },
          "terminal_outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal Outcome"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/WarningResponse"
            },
            "title": "Warnings",
            "type": "array"
          },
          "working_draft_state": {
            "$ref": "#/components/schemas/WorkingDraftState"
          }
        },
        "required": [
          "recruitment_status",
          "preparation_state",
          "working_draft_state",
          "review_reasons",
          "stale_reasons",
          "warnings",
          "active_job_snapshot_id",
          "newer_draft_in_progress",
          "available_actions",
          "blocked_actions",
          "application",
          "latest_snapshot"
        ],
        "title": "ApplicationDetailResponse",
        "type": "object"
      },
...
      "ApplicationListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApplicationListItemResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ApplicationListResponse",
        "type": "object"
      },
      "ApplicationMutationResponse": {
        "additionalProperties": false,
        "description": "One Application's tracked state after a write, and the event that moved it.",
        "properties": {
          "application_id": {
            "title": "Application Id",
            "type": "string"
          },
          "current_status": {
            "title": "Current Status",
            "type": "string"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "next_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Action"
          },
          "next_action_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Action Date"
          },
          "terminal_outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal Outcome"
          }
        },
        "required": [
          "application_id",
          "current_status"
        ],
        "title": "ApplicationMutationResponse",
        "type": "object"
      },
...
      "CorrectStatusRequest": {
        "additionalProperties": false,
        "description": "A reasoned correction appended to the trail.\n\nThe corrected event is never deleted or altered, so both the event being\ncorrected and the reason are required: a correction that named neither\nwould be indistinguishable from an ordinary transition.",
        "properties": {
          "corrects_event_id": {
            "minLength": 1,
            "title": "Corrects Event Id",
            "type": "string"
          },
          "occurred_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "target_status": {
            "enum": [
              "saved",
              "applied",
              "recruiter_screen",
              "interview",
              "assignment",
              "final_stage",
              "offer",
              "accepted",
              "rejected",
              "withdrawn",
              "closed"
            ],
            "title": "Target Status",
            "type": "string"
          }
        },
        "required": [
          "target_status",
          "corrects_event_id",
          "reason"
        ],
        "title": "CorrectStatusRequest",
        "type": "object"
      },
...
      "CreateApplicationRequest": {
        "additionalProperties": false,
        "properties": {
          "acknowledged_duplicates": {
            "default": false,
            "title": "Acknowledged Duplicates",
            "type": "boolean"
          },
          "company": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Company",
            "type": "string"
          },
          "job_text": {
            "minLength": 1,
            "title": "Job Text",
            "type": "string"
          },
          "source_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "target_role": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Target Role",
            "type": "string"
          }
        },
        "required": [
          "company",
          "target_role",
          "job_text"
        ],
        "title": "CreateApplicationRequest",
        "type": "object"
      },
      "CreateApplicationResponse": {
        "additionalProperties": false,
        "properties": {
          "application_id": {
            "title": "Application Id",
            "type": "string"
          },
          "duplicate_matches": {
            "items": {
              "$ref": "#/components/schemas/DuplicateMatchResponse"
            },
            "title": "Duplicate Matches",
            "type": "array"
          },
          "job_snapshot_id": {
            "title": "Job Snapshot Id",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "application_id",
          "job_snapshot_id",
          "warnings",
          "duplicate_matches"
        ],
        "title": "CreateApplicationResponse",
        "type": "object"
      },
...
      "ExternalSubmissionRequest": {
        "additionalProperties": false,
        "description": "A submission made outside the system, recorded without inventing evidence.\n\n`artifact_version_id` may name an already registered artifact, and stays\nabsent when there is none: a field that cannot be derived stays null rather\nthan being filled with a value the record never carried.",
        "properties": {
          "artifact_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Version Id"
          },
          "metadata": {
            "additionalProperties": true,
            "default": {},
            "title": "Metadata",
            "type": "object"
          },
          "submitted_at": {
            "minLength": 1,
            "title": "Submitted At",
            "type": "string"
          }
        },
        "required": [
          "submitted_at"
        ],
        "title": "ExternalSubmissionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NextActionRequest": {
        "additionalProperties": false,
        "description": "Set or clear the one active next action, stated in whole.\n\nBoth fields are required and both are nullable. Clearing is a real request\n- sending `null` for each is how a client says the action is done - and\nrequiring them is what keeps that distinguishable from an omitted field\nmeaning \"leave it alone\".",
        "properties": {
          "next_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Action"
          },
          "next_action_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Action Date"
          }
        },
        "required": [
          "next_action",
          "next_action_date"
        ],
        "title": "NextActionRequest",
        "type": "object"
      },
...
      "SubmissionResponse": {
        "additionalProperties": false,
        "description": "A recorded submission, with any staleness the caller should see.\n\n`warnings` carries the `READY_REVISION_FOR_OLDER_*` codes: the submission\nsucceeded, and the active snapshot or analysis has moved on since the\nrevision was approved. They are reported, not raised - active-context\ncompatibility is not a precondition for submitting a qualified revision.",
        "properties": {
          "application_id": {
            "title": "Application Id",
            "type": "string"
          },
          "approved_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved Revision Id"
          },
          "current_status": {
            "title": "Current Status",
            "type": "string"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "next_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Action"
          },
          "next_action_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Action Date"
          },
          "pdf_artifact_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Artifact Version Id"
          },
          "submission_id": {
            "title": "Submission Id",
            "type": "string"
          },
          "terminal_outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal Outcome"
          },
          "warnings": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "application_id",
          "current_status",
          "submission_id"
        ],
        "title": "SubmissionResponse",
        "type": "object"
      },
      "SubmitApplicationRequest": {
        "additionalProperties": false,
        "description": "One exact qualified revision and the exact PDF that was sent.\n\nBoth IDs are explicit. A submission that resolved the latest revision for\nitself could record having sent something the user never saw, which is the\none claim in this system that cannot be re-derived afterwards.",
        "properties": {
          "approved_revision_id": {
            "minLength": 1,
            "title": "Approved Revision Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "default": {},
            "title": "Metadata",
            "type": "object"
          },
          "pdf_artifact_version_id": {
            "minLength": 1,
            "title": "Pdf Artifact Version Id",
            "type": "string"
          },
          "submitted_at": {
            "minLength": 1,
            "title": "Submitted At",
            "type": "string"
          }
        },
        "required": [
          "approved_revision_id",
          "pdf_artifact_version_id",
          "submitted_at"
        ],
        "title": "SubmitApplicationRequest",
        "type": "object"
      },
...
      "TransitionStatusRequest": {
        "additionalProperties": false,
        "description": "An allowed forward transition. `reason` is optional for a normal move.",
        "properties": {
          "occurred_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "target_status": {
            "enum": [
              "saved",
              "recruiter_screen",
              "interview",
              "assignment",
              "final_stage",
              "offer",
              "accepted",
              "rejected",
              "withdrawn",
              "closed"
            ],
            "title": "Target Status",
            "type": "string"
          }
        },
        "required": [
          "target_status"
        ],
        "title": "TransitionStatusRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CV Engine local API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/applications": {
      "get": {
        "operationId": "list_applications_api_v1_applications_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List applications",
        "tags": [
          "applications"
        ]
      },
      "post": {
        "operationId": "create_application_api_v1_applications_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateApplicationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create an application and its first immutable job snapshot",
        "tags": [
          "applications"
        ]
      }
    },
...
    "/api/v1/applications/{application_id}": {
      "get": {
        "operationId": "application_detail_api_v1_applications__application_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read an application and its active preparation state",
        "tags": [
          "applications"
        ]
      }
    },
...
    "/api/v1/applications/{application_id}/external-submissions": {
      "post": {
        "description": "`201`: recorded without inventing a revision or an artifact that never was.",
        "operationId": "record_external_submission_api_v1_applications__application_id__external_submissions_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalSubmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record a submission made outside the system",
        "tags": [
          "tracking"
        ]
      }
    },
...
    "/api/v1/applications/{application_id}/next-action": {
      "patch": {
        "description": "States what the one active next action now is, in whole.\n\nSending both fields empty clears it: the request carries the complete value\nrather than a partial edit, and the schema requires both fields so an\nomitted one cannot be read as \"leave it alone\". `PATCH` rather than `PUT`\nbecause `PUT` is outside this API's allowed methods, and widening transport\npolicy for one route's verb would be the wrong trade.",
        "operationId": "set_next_action_api_v1_applications__application_id__next_action_patch",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NextActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set or clear the one active next action",
        "tags": [
          "tracking"
        ]
      }
    },
    "/api/v1/applications/{application_id}/status": {
      "post": {
        "description": "`200`: asking for the status an Application already holds is not an error.\n\nThe application layer returns the current state unchanged rather than\nappending a second identical event, so a client that retries a transition\nit already made does not write a duplicate into the trail.",
        "operationId": "transition_status_api_v1_applications__application_id__status_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transition recruitment status with an immutable history",
        "tags": [
          "tracking"
        ]
      }
    },
    "/api/v1/applications/{application_id}/status-corrections": {
      "post": {
        "description": "`201`: a correction appends an event; it never edits the one it corrects.",
        "operationId": "correct_status_api_v1_applications__application_id__status_corrections_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Append a reasoned correction to recruitment history",
        "tags": [
          "tracking"
        ]
      }
    },
    "/api/v1/applications/{application_id}/submissions": {
      "post": {
        "description": "`201`: an immutable Submission, refused unless the evidence still qualifies.\n\nReady qualification is re-derived from stored evidence at submission time,\nso a revision whose PDF was replaced or whose hashes no longer match is a\n`412` rather than a recorded claim that something was sent.",
        "operationId": "submit_application_api_v1_applications__application_id__submissions_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record an internal submission of one exact qualified revision",
        "tags": [
          "tracking"
        ]
      }
    },
...
  }
}